# src/test/resources/swagger/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OGC API - Processes",
    "description": "Example API Definition for OGC API - Processes",
    "contact": {
      "name": "Open Geospatial Consortium",
      "email": "info@ogc.org"
    },
    "license": {
      "name": "OGC License",
      "url": "http://www.opengeospatial.org/legal/"
    },
    "version": "0.1"
  },
  "servers": [
    {
      "url": "http://localhost/testogcapi",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/processes/{processId}": {
      "get": {
        "tags": [
          "DRU",
          "Processes"
        ],
        "summary": "retrieve a process description",
        "description": "The process description contains information about inputs and outputs and a link to the execution-endpoint for the process. The Core does not mandate the use of a specific process description to specify the interface of a process. That said, the Core requirements class makes the following recommendation:  Implementations SHOULD consider supporting the OGC process description.  For more information, see [Section 7.8](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_process_description). ",
        "operationId": "getProcessDescription",
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A process description.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Process"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "DRU"
        ],
        "summary": "replace a process.",
        "description": "Replaces a process.  For more information, see [Section 6.4](http://docs.ogc.org/DRAFTS/20-044.html#_18582f42-ebc6-4284-9333-c089068f62b6). ",
        "operationId": "replace",
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/ogcapppkg+json": {
              "schema": {
                "$ref": "#/components/schemas/Ogcapppkg"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "successful operation (no response body)"
          },
          "400": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "403": {
            "description": "the processes is not mutable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "409": {
            "description": "the processes being added is already deployed (i.e. duplicate)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "500": {
            "description": "A server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "DRU"
        ],
        "summary": "undeploy a process.",
        "description": "Undeploys a process.  For more information, see [Section 6.5](http://docs.ogc.org/DRAFTS/20-044.html#_16391f9e-538f-4a84-9710-72a6bab82842). ",
        "operationId": "undeploy",
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "successful operation (no response body)"
          },
          "403": {
            "description": "the processes is not mutable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "500": {
            "description": "A server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
    "/processes": {
      "get": {
        "tags": [
          "DRU",
          "Processes"
        ],
        "summary": "retrieve the list of available processes",
        "description": "The list of processes contains a summary of each process the OGC API - Processes offers, including the link to a more detailed description of the process.  For more information, see [Section 7.7]https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_process_list). ",
        "operationId": "getProcesses",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Information about the available processes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DRU"
        ],
        "summary": "deploy a process.",
        "description": "Deploys a process.  For more information, see [Section 6.3](http://docs.ogc.org/DRAFTS/20-044.html#_87a6983e-d060-458c-95ab-27e232e64822). ",
        "operationId": "deploy",
        "requestBody": {
          "content": {
            "application/ogcapppkg+json": {
              "schema": {
                "$ref": "#/components/schemas/Ogcapppkg"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Process Summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessSummary"
                }
              }
            }
          },
          "400": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "403": {
            "description": "the processes is not mutable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "409": {
            "description": "the processes being added is already deployed (i.e. duplicate)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "500": {
            "description": "A server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
...
    "/processes/{processId}/package": {
      "get": {
        "tags": [
          "DRU"
        ],
        "summary": "retrieve the formal description used to deploy a process",
        "description": "Access the formal description that can be used to deploy a process on an OGC API - Processes Server Instance. ",
        "operationId": "getPackage",
        "parameters": [
          {
            "name": "processId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The formal used to deploy a process",
            "content": {
              "application/ogcapppkg+json": {
                "schema": {
                  "$ref": "#/components/schemas/Ogcapppkg"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ogcapppkg"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/ogcapppkg+json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{jobId}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "retrieve the status of a job",
        "description": "Shows the status of a job.   For more information, see [Section 7.10](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_retrieve_status_info). ",
        "operationId": "getStatus",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "local identifier of a job",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The status of a job.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "500": {
            "description": "A server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "cancel a job execution, remove a finished job",
        "description": "Cancel a job execution and remove it from the jobs list.  For more information, see [Section 14]https://docs.ogc.org/is/18-062r2/18-062r2.html#Dismiss). ",
        "operationId": "dismiss",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "local identifier of a job",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The status of a job.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist on the server. For example, a path parameter had an incorrect value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          },
          "500": {
            "description": "A server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "exception": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        },
        "description": "JSON schema for exceptions based on RFC 7807"
      },
      "InputDescription": {
        "required": [
          "schema"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Metadata"
            }
          },
          "schema": {
            "$ref": "#/components/schemas/ProcessesCoreSchema"
          },
          "dataClasses": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "minOccurs": {
            "type": "integer",
            "format": "int32"
          },
          "maxOccurs": {
            "$ref": "#/components/schemas/InputDescriptionAllOfMaxOccurs"
          },
          "valuePassing": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "byValue",
                "byReference"
              ]
            }
          }
        }
      },
      "InputDescriptionAllOfMaxOccurs": {
        "type": "object"
      },
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Metadata": {
        "type": "object"
      },
      "Ogcapppkg": {
        "required": [
          "executionUnit"
        ],
        "type": "object",
        "properties": {
          "processDescription": {
            "$ref": "#/components/schemas/OgcapppkgProcessDescription"
          },
          "executionUnit": {
            "$ref": "#/components/schemas/OgcapppkgExecutionUnit"
          }
        }
      },
      "OgcapppkgExecutionUnit": {
        "type": "object"
      },
      "OgcapppkgProcessDescription": {
        "required": [
          "process"
        ],
        "type": "object",
        "properties": {
          "process": {
            "$ref": "#/components/schemas/Process"
          }
        }
      },
      "OutputDescription": {
        "required": [
          "schema"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Metadata"
            }
          },
          "schema": {
            "$ref": "#/components/schemas/ProcessesCoreSchema"
          }
        }
      },
      "Process": {
        "required": [
          "id",
          "version"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Metadata"
            }
          },
          "id": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "jobControlOptions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "sync-execute",
                "async-execute",
                "dismiss"
              ]
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          },
          "inputs": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/InputDescription"
            }
          },
          "outputs": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/OutputDescription"
            }
          }
        }
      },
      "ProcessesCoreSchema": {
        "type": "object"
      },
      "ProcessSummary": {
        "required": [
          "id",
          "version"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Metadata"
            }
          },
          "id": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "jobControlOptions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "sync-execute",
                "async-execute",
                "dismiss"
              ]
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          }
        }
      },
      "StatusInfo": {
        "required": [
          "id",
          "status",
          "type"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Metadata"
            }
          },
          "id": {
            "type": "string"
          },
          "processID": {
            "type": "string",
            "format": "uri"
          },
          "type": {
            "type": "string",
            "example": "[process, wps, openeo]"
          },
          "request": {
            "$ref": "#/components/schemas/StatusInfoAllOfRequest"
          },
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "running",
              "successful",
              "failed",
              "dismissed"
            ]
          },
          "message": {
            "type": "string"
          },
          "exception": {
            "$ref": "#/components/schemas/exception"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "started": {
            "type": "string",
            "format": "date-time"
          },
          "finished": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "progress": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          }
        }
      },
      "StatusInfoAllOfRequest": {
        "type": "object"
      },
...
      "ProcessList": {
        "required": [
          "links",
          "processes"
        ],
        "type": "object",
        "properties": {
          "processes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessSummary"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          }
        }
      },
...
    }
  }
}